+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/AwsDataset.kt
@@ -61,7 +61,8 @@ internal class AwsDataset(
                     launchTimeouts = JiraLaunchTimeouts(
                         offlineTimeout = Duration.ofMinutes(8),
                         initTimeout = Duration.ofMinutes(4),
-                        upgradeTimeout = Duration.ofMinutes(8)
+                        upgradeTimeout = Duration.ofMinutes(8),
+                        unresponsivenessTimeout = Duration.ofMinutes(4)
                     )
                 ),
                 computer = C5NineExtraLargeEphemeral()