+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/RemoteLocation.kt
@@ -30,7 +30,7 @@ data class RemoteLocation(val host: SshHost, val location: String) {
 
     fun move(destination: String, timeout: Duration): RemoteLocation {
         if (location != destination) {
-            Ssh(host).newConnection().use {
+            Ssh(host, connectivityPatience = 4).newConnection().use {
                 it.execute("mv $location $destination", timeout)
             }
         }
@@ -39,7 +39,7 @@ data class RemoteLocation(val host: SshHost, val location: String) {
 
     fun archive(timeout: Duration): RemoteLocation {
         logger.info("Archiving $location...")
-        val destination = Ssh(host).newConnection().use {
+        val destination = Ssh(host, connectivityPatience = 4).newConnection().use {
             FileArchiver().zip(it, location, timeout)
         }
         logger.info("Archiving $location complete")
@@ -48,14 +48,14 @@ data class RemoteLocation(val host: SshHost, val location: String) {
 
     fun upload(storage: StorageLocation, timeout: Duration) {
         logger.info("Uploading $location...")
-        Ssh(host).newConnection().use { AwsCli().uploadFile(storage, it, location, timeout) }
+        Ssh(host, connectivityPatience = 4).newConnection().use { AwsCli().uploadFile(storage, it, location, timeout) }
         logger.info("Uploading $location complete")
     }
 
     fun download(
         localDestination: Path
     ) {
-        Ssh(host).newConnection().use {
+        Ssh(host, connectivityPatience = 4).newConnection().use {
             it.download(
                 remoteSource = location,
                 localDestination = localDestination