+++ src/test/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/jira/StandaloneFormulaIT.kt
@@ -9,6 +9,8 @@ import com.atlassian.performance.tools.awsinfrastructure.api.hardware.C5NineExtr
 import com.atlassian.performance.tools.awsinfrastructure.api.storage.JiraServiceDeskStorage
 import com.atlassian.performance.tools.infrastructure.api.app.Apps
 import com.atlassian.performance.tools.infrastructure.api.app.NoApp
+import com.atlassian.performance.tools.infrastructure.api.jira.JiraJvmArgs
+import com.atlassian.performance.tools.infrastructure.api.jira.JiraLaunchTimeouts
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraNodeConfig
 import org.junit.Test
 import java.time.Duration
@@ -36,7 +38,16 @@ class StandaloneFormulaIT {
             application = JiraServiceDeskStorage("3.9.8"),
             database = dataset.database,
             jiraHomeSource = dataset.jiraHomeSource,
-            config = JiraNodeConfig(),
+            config = JiraNodeConfig(
+                name = "jira-node",
+                jvmArgs = JiraJvmArgs(),
+                launchTimeouts = JiraLaunchTimeouts(
+                    offlineTimeout = Duration.ofMinutes(8),
+                    initTimeout = Duration.ofMinutes(4),
+                    upgradeTimeout = Duration.ofMinutes(8),
+                    unresponsivenessTimeout = Duration.ofMinutes(4)
+                )
+            ),
             computer = C5NineExtraLargeEphemeral()
         )