+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 - Add support for other EC2 instance types. Resolve [JPERF-276].
 - Add support for `infrastructure:3`
 - Add support for `ssh:2`
+- Respect the `JiraLaunchTimeouts.unresponsivenessTimeout` in `StandaloneStoppedNode`. Resolve [JPERF-271]. 
 
 ### Fixed
 - Reduce apt-get usage in `AwsCli`. Mitigate [JPERF-219].
@@ -40,6 +41,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 [JPERF-219]: https://ecosystem.atlassian.net/browse/JPERF-219
 [JPERF-266]: https://ecosystem.atlassian.net/browse/JPERF-266
+[JPERF-271]: https://ecosystem.atlassian.net/browse/JPERF-271
 [JPERF-276]: https://ecosystem.atlassian.net/browse/JPERF-276
 
 ## [1.4.0] - 2018-10-31